requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-environ==0.9.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0; python_version >= '3.6'
django==4.1.3
pillow==9.3.0
psycopg2==2.9.5
sqlparse==0.4.3; python_version >= '3.5'
